glyb: fix compiler warnings for float and integer casts
- use size_t for buffer length variables to fix warnings
- add suffixes to floating point constants to fix warnings
- add floating point casts to suppress integer cast warnings

+++ src/logger.cc
@@ -25,16 +25,16 @@ void logger::output(const char *prefix, const char* fmt, va_list args1)
 {
     std::vector<char> buf;
     va_list args2;
-    int plen, pout, len, ret;
+    size_t plen, pout, len, ret;
 
     va_copy(args2, args1);
 
     plen = strlen(prefix);
-    len = vsnprintf(NULL, 0, fmt, args1);
+    len = (size_t)vsnprintf(NULL, 0, fmt, args1);
     assert(len >= 0);
     buf.resize(plen + len + 4); // prefix + ": " + message + CR + zero-terminator
-    pout = snprintf(buf.data(), buf.capacity(), "%s: ", prefix);
-    ret = vsnprintf(buf.data() + pout, buf.capacity(), fmt, args2);
+    pout = (size_t)snprintf(buf.data(), buf.capacity(), "%s: ", prefix);
+    ret = (size_t)vsnprintf(buf.data() + pout, buf.capacity(), fmt, args2);
     assert(len == ret);
     if (buf[buf.size()-3] != '\n') buf[buf.size()-2] = '\n';
